Big refactor to deal with alignment issues
The previous practice of casting bytes to uint64_t could
incur a performance penalty or cash if memory was not properly
aligned. Do all these operations through a union to ensure
proper alignment. Also change uint8_t to unsigned char since
compilers are not guaranteed to treat the former as a character
type for aliasing purposes.