Adds option to disable building docs.
It's useful to have a packaging option to determine whether to require
asciidoc and build the manpages.  But with the condition being solely
based on the opportunistic discovery of asciidoc, simply removing the
dependency gives non-reproducible results.  This adds an option to
explicitly disable the documentation, by pretending that 'a2x' doesn't
exist.

TESTED:
With asciidoc present, manpages are still built by default, but not
built if "-DDISABLE_DOCS=1 is included on the cmake command line.