Fixes test crash with OpenSSL < 1.0.2b.
Prior to OpenSSL 1.0.2b, CMAC_CTX_free() failed to check for a NULL
argument, resulting in a crash when so invoked.  This was caught in
the allocation-failure test.

This adds NULL checks to the two calls to CMAC_CTX_free() in
AES_SIV_CTX_free().  Although these could be made conditional on the
OpenSSL version, it's probably not worth it.

TESTED:
Previously crashing tests now pass on Ubuntu 14.04 with OpenSSL
1.0.1f.  They also pass on macOS 10.9.5 with OpenSSL 1.0.2q, CentOS 7
with OpenSSL 1.0.2k-fips, Fedora 25 with OpenSSL 1.0.2m-fips, and
FreeBSD 10.3 with OpenSSL 1.0.1s-freebsd.  Not tested on OpenBSD
or NetBSD due to cmake deficiencies.