Fixes out-of-source manpage builds.
The a2x program implicitly writes its outputs to the source directory.
In an out-of-source build, this means that the manpages are generated
in the source directory rather than the binary directory, and then the
install fails to find them.

In principle, the fix is just a matter of using a2x's -D option
(a.k.a. --destination-dir), but both the documentation and the code
believe (apparently incorrectly) that it is only effective for the
HTML output formats, and a warning is generated.  See:

     https://github.com/asciidoc/asciidoc/issues/44

To minimize the chatter, this fix only applies the option in cases
where it is actually necessary.

TESTED:
Ran build and install, both in-source and out-of source, with correct
results in both cases, and no warnings in the in-source case.